backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Scheduling
apscheduler==3.10.4

# Logging
structlog==23.2.0

# Retry Logic
tenacity==8.2.3

# CORS
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2
